requirements.txt
streamlit==1.36.0
openpyxl==3.1.5
reportlab==4.2.2
qrcode[pil]==7.4.2
pillow==10.4.0